[OVZC-55] 각 기능 패키지 분류 및 이전 코드 수정
