+++ src/main/java/kr/ac/kgu/kpserver/domain/board/Board.java
@@ -1,6 +1,7 @@
 package kr.ac.kgu.kpserver.domain.board;
 
 import kr.ac.kgu.kpserver.domain.BaseEntity;
+import kr.ac.kgu.kpserver.domain.board.Articles.Article;
 import lombok.AllArgsConstructor;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
@@ -27,7 +28,4 @@ public class Board extends BaseEntity {
 
     @OneToMany(mappedBy = "board")
     private List<Article> articles = new ArrayList<>();
-
-
-
 }
