+++ src/main/java/kr/ac/kgu/kpserver/domain/board/Comments/CommentRequest.java
@@ -1,0 +1,14 @@
+package kr.ac.kgu.kpserver.domain.board.Comments;
+
+import lombok.*;
+
+@Setter
+@Getter
+@NoArgsConstructor
+@AllArgsConstructor
+public class CommentRequest {
+    private Long userId;
+    private Long articleId;
+    private String description;
+
+}
