--- src/main/java/kr/ac/kgu/kpserver/domain/board/dto/LikeRequest.java
@@ -1,15 +1,0 @@
-package kr.ac.kgu.kpserver.domain.board.dto;
-
-import lombok.AccessLevel;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-import lombok.Setter;
-
-@Setter
-@Getter
-@NoArgsConstructor(access = AccessLevel.PROTECTED)
-public class LikeRequest {
-
-    private Long userId;
-    private Long articleId;
-}
